--- meeshkan/kafka/__init__.py
@@ -1,4 +1,0 @@
-from .processor import *
-from . import processor
-
-__all__ = processor.__all__
